Move addon related logic to pkg/addons for better organization
